+++ ta4j-core/src/main/java/org/ta4j/core/reports/package-info.java
@@ -21,41 +21,8 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.tradereport;
-
-import org.ta4j.core.num.Num;
-
 /**
- * This class represents report which contains performance statistics
+ * The main module for trade reports
+ *
  */
-public class PerformanceReport {
-
-    private final Num totalProfitLoss;
-    private final Num totalProfitLossPercentage;
-    private final Num totalProfit;
-    private final Num totalLoss;
-
-    public PerformanceReport(Num totalProfitLoss, Num totalProfitLossPercentage, Num totalProfit, Num totalLoss) {
-        this.totalProfitLoss = totalProfitLoss;
-        this.totalProfitLossPercentage = totalProfitLossPercentage;
-        this.totalProfit = totalProfit;
-        this.totalLoss = totalLoss;
-    }
-
-    public Num getTotalProfitLoss() {
-        return totalProfitLoss;
-    }
-
-    public Num getTotalProfitLossPercentage() {
-        return totalProfitLossPercentage;
-    }
-
-    public Num getTotalProfit() {
-        return totalProfit;
-    }
-
-    public Num getTotalLoss() {
-        return totalLoss;
-    }
-
-}
+package org.ta4j.core.reports;
